All actor refs now contain an UniqueNode, even local ones; this vastly simplifies many patterns and removes a lot of fatalErrors in our code; Eases SWIM integration since it assumes it is working with Nodes

+++ Tests/DistributedActorsTests/Cluster/Protobuf/Membership+SerializationTests.swift
@@ -51,7 +51,7 @@ final class MembershipSerializationTests: ActorSystemXCTestCase {
             Cluster.Member(
                 node: UniqueNode(
                     node: Node(systemName: "\(id)", host: "1.1.1.\(id)", port: 1111),
-                    nid: .init(UInt32("\(id)\(id)\(id)\(id)")!) // pretend a real-ish looking ID, but be easier to read
+                    nid: .init(UInt64("\(id)\(id)\(id)\(id)")!) // pretend a real-ish looking ID, but be easier to read
                 ),
                 status: .up
             )
